Copied in all updated tests from here including those in Utils folder: https://github.com/csinn/CSharp-From-Zero-To-Hero/tree/Chapter1/Homework/4/Tests/BootCamp.Chapter.Tests
For PromptString, PromptInt, and PromptFloat:
Changed Console.WriteLine to Write for error messages, added 'Environment.NewLine +' in front of error messages, added 'Console.Write(Environment.NewLine);' in front of all returns that did not have an error message.

All updated tests passed.

That was ridiculous that things Enviornment.NewLine was necessary instead of just using WriteLine, \n, or allowing returns to go to the next line.