Homework 6 - Debugging
Fix for the code is as follows:
Increased the Task Delay in the For Loop for "Aysnchromaguss.Mumble()" from 10 to 100.
A task delay of 10 from original code gives the incorrect passcode of "A maius minore a ad solis usque ab ortu disce ad occasum uno omnes".
The newly changed task delay of 100 gives the correct passcode of "A minore ad maius a solis ortu usque ad occasum ab uno disce omnes".
I got confirmatiion from @Antero360 in classroom Discord that the problem with the Task Delay was intended, and that the problem was to be solved through trial and error (just as I had done by guessing). There does not seem to be a way do debug for this problem.

For Part 1, the return values are as follows:
- First return value is "Debugger!"
- Second return value is "51"
- Third return value is "79"
- Last return value is "77"

For Part 2, the passcode is as follows:
- "A minore ad maius a solis ortu usque ad occasum ab uno disce omnes"

Result is as follows:
  YOU MAY PROCEED!
  --------------------------------
  From the smaller to the greater,
  From sunrise to sunset,
  From one, learn all.
  --------------------------------

Side note: Thank you for teaching us.  There is so much to learn from all of you.