Homework 7 - String: Part 1: Create a border with given padding around a string. For example:          Input: "Hello", 0            +-----+            |Hello|            +-----+
         Input: $"Hello{Environment.NewLine}World!", 0
           +------+
           |Hello |
           |World!|
           +------+

         Input: "Hello", 1
           +-------+
           |       |
           | Hello |
           |       |
           +-------+

https://github.com/csinn/CSharp-From-Zero-To-Hero/blob/Chapter1/Homework/7/Src/BootCamp.Chapter/TextTable.cs

---

Part 2:
You have an array of strings. Each element in array (a string) follows the same structure:
Name1, balanceX1, balanceX2, balanceX3...
Name2, balanceY1, balanceY2, balanceY3...

So each line is a person with their balance history. The last balance in the array is their current amount of money they have. Currency in the balances is unspecified.

You need to parse the balances and process them in the following ways:
            // - FindHighestBalanceEver- Return name and balance(current) of person who had the biggest historic balance.
            // - FindPersonWithBiggestLoss- Return name and loss of a person with a biggest loss (balance change negative).
            // - FindRichestPerson- Return name and current money of the richest person.
            // - FindMostPoorPerson- Return name and current money of the most poor person.

Results of each operation have to be printed using the TextTableBuilder, using padding = 3 (so 4 calls to the TextTableBuilder).

Note: People and their balances can be found in: PeopleBalances.cs, Blaances property.